backend/requirements.txt
celery==5.3.0
pymongo==3.12.0
google-api-python-client==2.35.0
python-dotenv==0.19.2
